+++ java/blogPessoal/src/main/java/com/org/generation/blogPessoal/controller/TemaController.java
@@ -1,6 +1,9 @@
 package com.org.generation.blogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
+
+import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -13,7 +16,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.org.generation.blogPessoal.model.TemaModel;
 import com.org.generation.blogPessoal.repository.TemaRepository;
@@ -42,18 +47,23 @@ public class TemaController {
 	}
 
 	@PostMapping
-	public ResponseEntity<TemaModel> post(@RequestBody TemaModel tema) {
+	public ResponseEntity<TemaModel> post(@Valid @RequestBody TemaModel tema) {
 		return ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema));
 	}
 
 	@PutMapping
-	public ResponseEntity<TemaModel> put(@RequestBody TemaModel tema) {
-		return ResponseEntity.ok(repository.save(tema));
+	public ResponseEntity<TemaModel> put(@Valid @RequestBody TemaModel tema){
+	        return repository.findById(tema.getId())
+			        .map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+			        .orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
+	@ResponseStatus(HttpStatus.NO_CONTENT)
 	@DeleteMapping("/{id}")
 	public void delete(@PathVariable long id) {
-		repository.deleteById(id);
-	}
-
+	        Optional<TemaModel> tema = repository.findById(id);
+	        if(tema.isEmpty())
+		        throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+	        repository.deleteById(id);
+  }
 }
